--- packages/ui/src/components/vuestic-components/va-avatar/VaAvatar.spec.js
@@ -1,9 +1,0 @@
-import VaAvatar from './VaAvatar'
-
-import { testIsLoadingMixin } from '../../vuestic-mixins/LoadingMixin/testIsLoadingMixin'
-
-describe('VaAvatar', () => {
-  it('has loading mixin', () => {
-    expect(() => testIsLoadingMixin(VaAvatar)).not.toThrow()
-  })
-})
