+++ packages/ui/src/components/vuestic-components/va-breadcrumbs/tests/VaBreadcrumbs.spec.ts
@@ -1,0 +1,10 @@
+import { mount } from '@vue/test-utils'
+// @ts-ignore
+import VaBreadcrumbs from '../VaBreadcrumbs.vue'
+
+describe('VaBreadcrumbs', () => {
+  it('should render without an error', () => {
+    const wrapper = mount(VaBreadcrumbs)
+    expect(wrapper.isVueInstance()).toBeTruthy()
+  })
+})
