+++ packages/ui/src/components/vuestic-components/va-card/tests/VaCard.spec.ts
@@ -1,7 +1,14 @@
-import VaCard from './VaCard'
-import { testIsContextableComponent } from '../../context-test/context-provide/testIsContextableComponent'
+import { mount } from '@vue/test-utils'
+// @ts-ignore
+import VaCard from '../VaCard.vue'
+
+import { testIsContextableComponent } from '../../../context-test/context-provide/testIsContextableComponent'
 
 describe('VaCard', () => {
+  it('should render without an error', () => {
+    const wrapper = mount(VaCard)
+    expect(wrapper.isVueInstance()).toBeTruthy()
+  })
   it('is contextable', () => {
     const props = {
       tag: 'a',
