+++ packages/ui/src/components/vuestic-components/va-modal/tests/VaModal.spec.ts
@@ -1,10 +1,11 @@
 import noop from 'lodash/noop'
 import { mount } from '@vue/test-utils'
 
-import { testHasStatefulMixin } from '../../vuestic-mixins/StatefullMixin/testHasStatefulMixin'
-import { testIsContextableComponent } from '../../context-test/context-provide/testIsContextableComponent'
-import { StatefulMixin } from '../../vuestic-mixins/StatefullMixin/StatefulMixin'
-import VaModal from './VaModal.vue'
+import { testHasStatefulMixin } from '../../../vuestic-mixins/StatefullMixin/testHasStatefulMixin'
+import { testIsContextableComponent } from '../../../context-test/context-provide/testIsContextableComponent'
+import { StatefulMixin } from '../../../vuestic-mixins/StatefullMixin/StatefulMixin'
+// @ts-ignore
+import VaModal from '../VaModal.vue'
 
 describe('VaModal', () => {
   describe('stateful mixin', () => {
