+++ packages/ui/src/components/vuestic-components/va-progress-bar/progress-types/tests/VaProgressBar.spec.ts
@@ -1,0 +1,11 @@
+import { mount } from '@vue/test-utils'
+
+// @ts-ignore
+import VaProgressBar from '../VaProgressBar.vue'
+
+describe('VaProgressBar', () => {
+  it('should render without an error', () => {
+    const wrapper = mount(VaProgressBar)
+    expect(wrapper.isVueInstance()).toBeTruthy()
+  })
+})
