+++ packages/ui/src/components/vuestic-components/va-radio/tests/VaRadio.spec.ts
@@ -1,0 +1,11 @@
+import { mount } from '@vue/test-utils'
+
+// @ts-ignore
+import VaRadio from '../VaRadio.vue'
+
+describe('VaRadio', () => {
+  it('should render without an error', () => {
+    const wrapper = mount(VaRadio)
+    expect(wrapper.isVueInstance()).toBeTruthy()
+  })
+})
