+++ packages/ui/src/components/vuestic-components/va-slider/tests/VaSlider.spec.ts
@@ -1,0 +1,11 @@
+import { mount } from '@vue/test-utils'
+
+// @ts-ignore
+import VaSlider from '../VaSlider.vue'
+
+describe('VaSlider', () => {
+  it('should render without an error', () => {
+    const wrapper = mount(VaSlider)
+    expect(wrapper.isVueInstance()).toBeTruthy()
+  })
+})
